+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Option to listen on a socket instead of HOST:PORT.
+- Vision tool to change active tokens.
 
 ## [0.15.1] - 2019-05-15